updated linting
